:construction: wip: add test and separate test file
