+++ src/filters/get_task_instance.rs
@@ -28,7 +28,7 @@ pub struct Train {
     pub times: i32,
 }
 
-pub async fn handler(extract: Extract, db: util::Db) -> Result<impl warp::Reply, warp::Rejection> {
+pub async fn handler(extract: Extract, db: util::Db) -> Result<Reply, warp::Rejection> {
     let db = db.lock().await;
 
     let user = sqlx::query!("SELECT id FROM usr WHERE token = $1", extract.token)
@@ -70,7 +70,7 @@ pub async fn handler(extract: Extract, db: util::Db) -> Result<impl warp::Reply,
         progress: task.progress,
     };
 
-    Ok(warp::reply::json(&reply))
+    Ok(reply)
 }
 
 pub fn filter(
@@ -81,4 +81,5 @@ pub fn filter(
         .and(util::json_body())
         .and(util::with_db(db))
         .and_then(handler)
+        .map(|reply| warp::reply::json(&reply))
 }
