+++ src/filters/get_training_results.rs
@@ -19,7 +19,7 @@ pub struct Reply {
     pub done_at: chrono::NaiveDateTime,
 }
 
-pub async fn handler(extract: Extract, db: util::Db) -> Result<impl warp::Reply, warp::Rejection> {
+pub async fn handler(extract: Extract, db: util::Db) -> Result<Vec<Reply>, warp::Rejection> {
     let db = db.lock().await;
 
     let user = sqlx::query!("SELECT id FROM usr WHERE token = $1", extract.token)
@@ -41,7 +41,7 @@ pub async fn handler(extract: Extract, db: util::Db) -> Result<impl warp::Reply,
     .await
     .map_err(|_| util::ErrorMessage::new("failed to get training results"))?;
 
-    Ok(warp::reply::json(&reply))
+    Ok(reply)
 }
 
 pub fn filter(
@@ -52,4 +52,5 @@ pub fn filter(
         .and(util::json_body())
         .and(util::with_db(db))
         .and_then(handler)
+        .map(|reply| warp::reply::json(&reply))
 }
