+++ src/filters/proceed_task_instance.rs
@@ -9,7 +9,7 @@ pub struct Extract {
     pub progress: i32,
 }
 
-pub async fn handler(extract: Extract, db: util::Db) -> Result<impl warp::Reply, warp::Rejection> {
+pub async fn handler(extract: Extract, db: util::Db) -> Result<(), warp::Rejection> {
     let db = db.lock().await;
 
     let user = sqlx::query!("SELECT id FROM usr WHERE token = $1", extract.token)
@@ -28,7 +28,7 @@ pub async fn handler(extract: Extract, db: util::Db) -> Result<impl warp::Reply,
     .await
     .map_err(|_| util::ErrorMessage::new("failed to update a task instance"))?;
 
-    Ok(warp::http::StatusCode::OK)
+    Ok(())
 }
 
 pub fn filter(
@@ -39,4 +39,5 @@ pub fn filter(
         .and(util::json_body())
         .and(util::with_db(db))
         .and_then(handler)
+        .map(|_| warp::http::StatusCode::OK)
 }
